Modified committee page to have cards for each committee member
By creating a page for each committee member with the title of their name and their role we can use the standard listing structure of blowfish to organise each years committee members. I also modified thee pagination to 12 to allow for all the the committee members from this year to fit on one page. On each committee member page I choose to not show the date so that it doesn't show in the page preview, also I have used the date as a proxy for rank order, by giving the highest ranking members the most recent date they are displayed first. Currently the committee homepage is still based on the basic list layout page but this might be changed in the future to allow more editing.